requirements.txt
click==8.1.3
pytorch-ignite~=0.4.9
tensorboard==2.10.1
scipy==1.9.3
numpy==1.22.4
timm~=0.4.12
tqdm==4.49.0
